This will make it much less confusing for beginning users.
I made it so that passing the variable
TARGET=nameofthemodel
(i.e. without any .mdl) will cause only that mdl file to be compiled. Now the GUI needs to be modified so that it also passes the TARGET= variable.

(Also, now the output from Synth-A-Modeler is waaaaaaay less verbose and the old text gets written to a temporary file SAM-debug-compilation.txt.)



Also the file Makefile.HOWTO should be helpful for other users who need help adapting their own/future Faust makefiles to work with Synth-A-Modeler.