Made the new model pluck_one_string_pitch_control.mdl.
Also changed simpleStringTerm() so that its second argument was the cutoff frequency for the lowpass filter instead of the length (in samples) of the lowpass filter.  Because of this, I had to retroactively edit a few of the older models in order to make them compatible with the new syntax.